+++ _background.h
@@ -13,9 +13,13 @@ class Acolyte : public _background{
   public:
     void addLanguages(vector<string> *languages){
       string lang;
-      for (auto it = allLangs.cbegin(); it != allLangs.cend(); ++it)
-        std::cout << ' ' << *it;
-      std::cout << ": " << endl;
+      cout << "[";
+      for (auto it = allLangs.cbegin(); it != allLangs.cend(); ++it){
+        std::cout << *it;
+        if(*it != allLangs.back())
+          cout << ", ";
+      }
+      cout << "]: ";
       std::cin >> lang;
       (*languages).push_back(lang);
       allLangs.remove(lang);
@@ -33,9 +37,13 @@ class Sage : public _background{
   public:
     void addLanguages(vector<string> *languages){
       string lang;
-      for (auto it = allLangs.cbegin(); it != allLangs.cend(); ++it)
-        std::cout << ' ' << *it;
-      std::cout << ": " << endl;
+      cout << "[";
+      for (auto it = allLangs.cbegin(); it != allLangs.cend(); ++it){
+        std::cout << *it;
+        if(*it != allLangs.back())
+          cout << ", ";
+      }
+      cout << "]: ";
       std::cin >> lang;
       (*languages).push_back(lang);
       allLangs.remove(lang);
@@ -53,10 +61,20 @@ class HauntedOne : public _background{
   public:
     void addLanguages(vector<string> *languages){
       string lang;
-      for (auto it = allLangs.cbegin(); it != allLangs.cend(); ++it)
-        std::cout << ' ' << *it;
-      std::cout << ": " << endl;
-      std::cin >> lang;
+      bool c = false;
+      do{
+        cout << "[";
+        for (int i = 0; i < (sizeof(exotics)/sizeof(*exotics)); i++){
+          std::cout << exotics[i];
+          if(i != 7)
+            cout << ", ";
+        }
+        cout << "]: ";
+        std::cin >> lang;
+        for(int i = 0; i < (sizeof(exotics)/sizeof(*exotics)); i++)
+          if(boost::iequals(lang, exotics[i]))
+            c = true;
+      }while(!c);
       (*languages).push_back(lang);
       allLangs.remove(lang);
     }
@@ -66,9 +84,13 @@ class Noble : public _background{
   public:
     void addLanguages(vector<string> *languages){
       string lang;
-      for (auto it = allLangs.cbegin(); it != allLangs.cend(); ++it)
-        std::cout << ' ' << *it;
-      std::cout << ": " << endl;
+      cout << "[";
+      for (auto it = allLangs.cbegin(); it != allLangs.cend(); ++it){
+        std::cout << *it;
+        if(*it != allLangs.back())
+          cout << ", ";
+      }
+      cout << "]: ";
       std::cin >> lang;
       (*languages).push_back(lang);
       allLangs.remove(lang);