+++ CHANGE LOG.md
@@ -4,7 +4,10 @@ Updated: 03/06/2020
 0.1.0: Auto rolls stats, organizes stats for class, sets save roll stats, sets armor and weapon proficiencies, sets hit die, boosts stats based on race, sets starting health.
 0.1.1: Saves are displayed as the stat name rather than their array location. Languages are assigned and extra languages can be choosen when appropriate.
 0.2.0: Refactored code.
+0.2.1: Race, Class, and Background selection are no longer case-sensitive. Bard can now choose 3 Instruments. Monk can now choose an Artisan's Tool or an Instrument. Stats output now shows the order of stats.
 
 Known Issues:
-Proficiency assignment doesn't work for Bard due to instrument selection not working yet.
-Proficiency assignment doesn't work for Monk due to not allowing selection between artisan tool or instrument and the selections that will come after.
+Instrument list doesn't always output correctly for Bard after making an instrument selection
+Skills are not assigned
+Languages for backgrounds that get languages (with the exception of Haunted One) will accept invalid input
+Languages provided from race are valid options when selecting background languages